requirements.txt
pypdf==4.0.0
PyPDF2==3.0.1
pandas==2.2.0
chromadb==0.5.0
google-generativeai==0.5.2
pysqlite3-binary==0.5.2.post3
